[carla] Aaron/upgradenurec #9175
